fix: Workaround for wrapper incompatibility
The combination of Gym's GrayScaleObservation and FrameStack wrappers
was incompatible to how stable_baselines3 expects the observation's
shape. A 3D-observation is assumed to be an RGB pixel image which does
not fit for a grayscale image (2D) composed into a frame stack (+ 1D).
A dummy dimension added to the gray scale observation is a workaround
for this issue.